Event listener added to specific modules removed from here.
