+++ api/pom.xml
@@ -45,20 +45,6 @@
         </dependency>
         <dependency>
             <groupId>org.bahmni.module</groupId>
-            <artifactId>appointments-api</artifactId>
-            <version>2.0.0-SNAPSHOT</version>
-            <type>jar</type>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.bahmni.module</groupId>
-            <artifactId>appointments-omod</artifactId>
-            <version>2.0.0-SNAPSHOT</version>
-            <type>jar</type>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.bahmni.module</groupId>
             <artifactId>bahmni-events-api</artifactId>
             <version>1.0.0-SNAPSHOT</version>
             <type>jar</type>
