--- api/src/main/java/org/bahmni/module/communication/scheduler/tasks/SMSReminderForAppointment.java
@@ -1,44 +1,0 @@
-package org.bahmni.module.communication.scheduler.tasks;
-
-
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-import org.bahmni.module.communication.service.SMSSenderService;
-import org.openmrs.PersonAttribute;
-import org.openmrs.api.AdministrationService;
-import org.openmrs.api.context.Context;
-import org.openmrs.module.appointments.model.Appointment;
-import org.openmrs.module.appointments.service.AppointmentsService;
-import org.openmrs.scheduler.tasks.AbstractTask;
-
-import java.util.List;
-
-public class SMSReminderForAppointment extends AbstractTask {
-    private Log log = LogFactory.getLog(this.getClass());
-
-
-    @Override
-    public void execute() {
-        AdministrationService administrationService = Context.getService(AdministrationService.class);
-
-        boolean scheduleSMS = Boolean.valueOf(administrationService.getGlobalPropertyObject("sms.enableAppointmentReminderSMSAlert").getPropertyValue());
-
-        if (!scheduleSMS) {
-            return;
-        }
-        AppointmentsService appointmentsService = Context.getService(AppointmentsService.class);
-        String schedulerReminderTime = administrationService.getGlobalPropertyObject("SchedulerReminderBeforeHours").getPropertyValue();
-        List<Appointment> appointments = appointmentsService.getAllAppointmentsReminder(schedulerReminderTime);
-        for (Appointment appointment: appointments) {
-            PersonAttribute phoneNumber = appointment.getPatient().getAttribute("phoneNumber");
-            if (null == phoneNumber) {
-                log.info("Since no mobile number found for the patient. SMS not sent.");
-                return;
-            }
-            SMSSenderService smsService =Context.getService(SMSSenderService.class);
-            String message = smsService.getAppointmentReminderMessage(appointment);
-            log.info("SMS response: " + smsService.sendSMS(phoneNumber.getValue(), message));
-        }
-        }
-
-}
