+++ api/src/main/java/org/bahmni/module/communication/service/CommunicationService.java
@@ -7,5 +7,7 @@ public interface CommunicationService {
 	
 	@Authorized({ "Send Email" })
 	void sendEmail(MailContent mailContent);
+
+	void sendSMS(String phoneNumber, String message);
 	
 }
