+++ api/src/main/resources/liquibase.xml
@@ -5,35 +5,4 @@
                    xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
          http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-2.0.xsd">
 
-    <changeSet id="drop-appointment-status-change-task-202313092962" author="Bahmni">
-        <preConditions onFail="MARK_RAN">
-            <!-- Check if the task exists -->
-            <sqlCheck expectedResult="1">
-                SELECT COUNT(*) FROM scheduler_task_config WHERE name ="Reminder of scheduled appointment";
-            </sqlCheck>
-        </preConditions>
-        <sql>
-            DELETE
-            FROM scheduler_task_config
-            WHERE name = 'Reminder of scheduled appointment';
-        </sql>
-    </changeSet>
-    <changeSet id="create-appointment-status-change-task-202311092962" author="Bahmni">
-        <preConditions>
-            <sqlCheck expectedResult="0">
-                SELECT COUNT(*)
-                FROM scheduler_task_config
-                WHERE name = "Reminder of scheduled appointment";
-            </sqlCheck>
-        </preConditions>
-        <sql>
-            INSERT INTO scheduler_task_config(name, description, schedulable_class, start_time, start_time_pattern,
-                                              repeat_interval,
-                                              start_on_startup, created_by, date_created, uuid)
-            VALUES ('Reminder of scheduled appointment', 'Reminder of scheduled appointment',
-                    'org.bahmni.module.communication.scheduler.tasks.SMSReminderForAppointment',
-                    TIMESTAMP (CURDATE()), 'MM/dd/yyyy HH:mm:ss', 3600, TRUE, 1, NOW(), UUID());
-        </sql>
-    </changeSet>
-
 </databaseChangeLog>
