+++ pom.xml
@@ -152,23 +152,9 @@
                 <scope>provided</scope>
             </dependency>
             <dependency>
-                <groupId>org.bahmni.module</groupId>
-                <artifactId>appointments-api</artifactId>
-                <version>2.0.0-SNAPSHOT</version>
-                <type>jar</type>
-                <scope>provided</scope>
-            </dependency>
-            <dependency>
                 <groupId>org.projectlombok</groupId>
                 <artifactId>lombok</artifactId>
                 <version>1.18.20</version>
-                <scope>provided</scope>
-            </dependency>
-            <dependency>
-                <groupId>org.bahmni.module</groupId>
-                <artifactId>appointments-omod</artifactId>
-                <version>2.0.0-SNAPSHOT</version>
-                <type>jar</type>
                 <scope>provided</scope>
             </dependency>
             <dependency>
